Add LoggerFactory(move getLogger method)
